feat(telemetry): add is_self_hosted tag
Add a boolean tag to Sentry telemetry to distinguish between SaaS
and self-hosted Sentry instances.

Also consolidates URL utilities:
- Move isSentrySaasUrl to sentry-urls.ts (canonical location)
- Remove duplicate DEFAULT_SENTRY_URL from sentry-urls.ts (use constants.ts)
- Remove getDefaultBaseUrl and isSentrySaasUrl from region.ts (use sentry-urls.ts)
